move team member task icons info to own component
